artifacts-service/pyproject.toml
[project]
name = "artifacts-service"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiobotocore>=3.0.0",
    "alembic>=1.14.0",
    "fastapi[standard]>=0.127.0",
    "psycopg[binary,pool]>=3.3.2",
    "pydantic-settings>=2.12.0",
    "sqlalchemy[asyncio,postgresql-asyncpg]>=2.0.45",
    "email-validator==2.3.0",
    "python-jose[cryptography]>=3.5.0",
    "httpx>=0.28.1",
    "reportlab>=4.4.7",
]
